+++ public/app/core/components/Page/Page.tsx
@@ -1,30 +1,19 @@
 // Libraries
 import { css, cx } from '@emotion/css';
-import React, { FC, HTMLAttributes } from 'react';
+import React from 'react';
 
-import { GrafanaTheme2, NavModel, NavModelItem } from '@grafana/data';
+import { GrafanaTheme2 } from '@grafana/data';
 import { config } from '@grafana/runtime';
 import { CustomScrollbar, useStyles2 } from '@grafana/ui';
 
 import { Footer } from '../Footer/Footer';
 import { PageHeader } from '../PageHeader/PageHeader';
-import { NewPage } from '../PageLayouts/NewPage';
+import { Page as NewPage } from '../PageNew/Page';
 
 import { PageContents } from './PageContents';
+import { PageType } from './types';
 import { usePageNav } from './usePageNav';
 import { usePageTitle } from './usePageTitle';
-
-interface Props extends HTMLAttributes<HTMLDivElement> {
-  children: React.ReactNode;
-  navId?: string;
-  navModel?: NavModel;
-  pageNav?: NavModelItem;
-}
-
-export interface PageType extends FC<Props> {
-  Header: typeof PageHeader;
-  Contents: typeof PageContents;
-}
 
 export const OldPage: PageType = ({ navId, navModel: oldNavProp, pageNav, children, className, ...otherProps }) => {
   const styles = useStyles2(getStyles);
